delete class and files repair tests
